+++ src/main/java/ch/nyp/xinnyi/bot/telegram/updatemodels/Entity.java
@@ -1,15 +1,15 @@
 package ch.nyp.xinnyi.bot.telegram.updatemodels;
 
 public class Entity {
-    private long offset;
-    private long length;
+    private int offset;
+    private int length;
     private String type;
 
     public long getOffset() {
         return offset;
     }
 
-    public Entity setOffset(long offset) {
+    public Entity setOffset(int offset) {
         this.offset = offset;
         return this;
     }
@@ -18,7 +18,7 @@ public class Entity {
         return length;
     }
 
-    public Entity setLength(long length) {
+    public Entity setLength(int length) {
         this.length = length;
         return this;
     }
